requirements.txt
aiogram==3.3.0
pytest==8.0.0
pytest-asyncio==0.23.3
python-dotenv==1.0.0